==> pgc1a/2.inspect-gene-compound.Rmd <==
---
title: "Inspect gene-compound connections"
output: html_notebook
---

```{r message=FALSE}
library(magrittr)
library(glue)
library(tidyverse)
```


This file was very likely created using this code
https://github.com/broadinstitute/2016_08_22_Gene_Compound/blob/f08ccc2b4763376dd00010bed559f3d531b61dd2/archive/CDRP_per_gene_analysis_mutants.Rmd#L335-L339

i.e. it has the Pearson correlation between genes and compounds

```{r}
connections <- readRDS("../archive/rds_files/cr_melt_cp.rds")
```


```{r}
set.seed(42)
connections %>%
  sample_n(10)
```


```{r}
connections %>%
  filter(Var2 == "PPARGC1A_WT.2") %>%
  count(name = "Number of connections with compounds")
```


```{r}
connections %>%
  filter(Var2 == "PPARGC1A_WT.2") %>%
  ggplot(aes(value)) + geom_histogram(bins = 100) +
  ggtitle("Distribution of Pearson similarity of compounds to PPARGC1A_WT.2")
```

We need `PUBCHEM_CID` from the `Var1` column

`Var1` is `CPD_NAME`, which was loaded here

https://github.com/broadinstitute/2016_08_22_Gene_Compound/blob/f08ccc2b4763376dd00010bed559f3d531b61dd2/archive/CDRP_per_gene_analysis_mutants.Rmd#L101

Let's load the CDRP metadata

```{r}
cdrp_cpd_metadata <-
  read_csv("input/CDP2/cdrp.cpd.meta.csv") %>%
  mutate(CPD_NAME = tolower(CPD_NAME))
```


```{r}
set.seed(42)
cdrp_cpd_metadata %>%
  sample_n(10)
```


```{r}
connections <-
  connections %>%
  left_join(
    connections %>%
      select(Var1) %>%
      inner_join(
        cdrp_cpd_metadata %>%
          select(CPD_NAME, BROAD_CPD_ID),
        by = c("Var1" = "CPD_NAME")
      ) %>%
      distinct()
  ) %>%
  mutate(BROAD_CPD_ID =
           ifelse(is.na(BROAD_CPD_ID),
                  str_sub(Var1, 1, 13),
                  BROAD_CPD_ID)) %>%
  inner_join(cdrp_cpd_metadata %>% select(BROAD_CPD_ID, cpd_type, CBP_SMILES))

```

We now have `BROAD_CPD_ID` for all compounds

```{r}
set.seed(42)
connections %>%
  sample_n(10)
```


```{r}
connections %>%
  distinct(BROAD_CPD_ID, cpd_type) %>%
  count(cpd_type)
```


```{r}
connections %>%
  filter(Var2 == "PPARGC1A_WT.2") %>%
  select(BROAD_CPD_ID, CBP_SMILES, value) %>%
  write_csv("input/PPARGC1A_WT_2_compound_connections.csv.gz")
```
